Install @babel/plugin-transform-modules-commonjs for ES2015 module support
I expect less-experienced JavaScript developers to be more familiar with
ES2015 modules (as seen in the browser) than with CommonJS modules (as
seen in Node.js), so I think it’s worth adding a few dependencies for
the sake of familiarity.

For similar reasons, I’m configuring Babel in `package.json` (rather
than e.g. `babel.config.js`) because it’s already present and
recognisable to most JavaScript developers.